requirements.txt
pytesseract>=0.3.10
Pillow>=9.5.0
opencv-python>=4.8.0
pymupdf>=1.23.26
python-docx>=1.1.0 
fastapi>=0.95.2
uvicorn>=0.22.0
pytest>=7.0.0
requests>=2.31.0
pdf2image>=1.16.0
httpx>=0.24.0